requirements.txt
Flask==2.0.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1